+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * `RoundedBackgroundTextField` now scales down to fit the text
 * `RoundedBackgroundTextField` now can't be scrollable
 * `RoundedBackgroundTextField` now uses `EditableText` instead of `TextField` in order to get the best positioning
+* Expose more text field properties
 
 ## 0.1.1