+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [unreleased]
+## 0.1.2
 
 * `RoundedBackgroundTextField` now scales down to fit the text
 * `RoundedBackgroundTextField` now can't be scrollable